Deduplicate Random class, provide serialVersionUID to KotlinRandom
