+++ libraries/stdlib/src/kotlin/random/Random.kt
@@ -5,6 +5,8 @@
 
 package kotlin.random
 
+import kotlin.math.nextDown
+
 /**
  * An abstract class that is implemented by random number generator algorithms.
  *
@@ -15,7 +17,7 @@ package kotlin.random
  * @sample samples.random.Randoms.defaultRandom
  */
 @SinceKotlin("1.3")
-public expect abstract class Random constructor() {
+public abstract class Random {
 
     /**
      * Gets the next random [bitCount] number of bits.
@@ -35,7 +37,7 @@ public expect abstract class Random constructor() {
      *
      * @sample samples.random.Randoms.nextInt
      */
-    public open fun nextInt(): Int
+    public open fun nextInt(): Int = nextBits(32)
 
     /**
      * Gets the next random non-negative `Int` from the random number generator less than the specified [until] bound.
@@ -48,7 +50,7 @@ public expect abstract class Random constructor() {
      *
      * @sample samples.random.Randoms.nextIntFromUntil
      */
-    public open fun nextInt(until: Int): Int
+    public open fun nextInt(until: Int): Int = nextInt(0, until)
 
     /**
      * Gets the next random `Int` from the random number generator in the specified range.
@@ -59,7 +61,29 @@ public expect abstract class Random constructor() {
      *
      * @sample samples.random.Randoms.nextIntFromUntil
      */
-    public open fun nextInt(from: Int, until: Int): Int
+    public open fun nextInt(from: Int, until: Int): Int {
+        checkRangeBounds(from, until)
+        val n = until - from
+        if (n > 0 || n == Int.MIN_VALUE) {
+            val rnd = if (n and -n == n) {
+                val bitCount = fastLog2(n)
+                nextBits(bitCount)
+            } else {
+                var v: Int
+                do {
+                    val bits = nextInt().ushr(1)
+                    v = bits % n
+                } while (bits - v + (n - 1) < 0)
+                v
+            }
+            return from + rnd
+        } else {
+            while (true) {
+                val rnd = nextInt()
+                if (rnd in from until until) return rnd
+            }
+        }
+    }
 
     /**
      * Gets the next random `Long` from the random number generator.
@@ -68,7 +92,7 @@ public expect abstract class Random constructor() {
      *
      * @sample samples.random.Randoms.nextLong
      */
-    public open fun nextLong(): Long
+    public open fun nextLong(): Long = nextInt().toLong().shl(32) + nextInt()
 
     /**
      * Gets the next random non-negative `Long` from the random number generator less than the specified [until] bound.
@@ -81,7 +105,7 @@ public expect abstract class Random constructor() {
      *
      * @sample samples.random.Randoms.nextLongFromUntil
      */
-    public open fun nextLong(until: Long): Long
+    public open fun nextLong(until: Long): Long = nextLong(0, until)
 
     /**
      * Gets the next random `Long` from the random number generator in the specified range.
@@ -92,21 +116,58 @@ public expect abstract class Random constructor() {
      *
      * @sample samples.random.Randoms.nextLongFromUntil
      */
-    public open fun nextLong(from: Long, until: Long): Long
+    public open fun nextLong(from: Long, until: Long): Long {
+        checkRangeBounds(from, until)
+        val n = until - from
+        if (n > 0) {
+            val rnd: Long
+            if (n and -n == n) {
+                val nLow = n.toInt()
+                val nHigh = (n ushr 32).toInt()
+                rnd = when {
+                    nLow != 0 -> {
+                        val bitCount = fastLog2(nLow)
+                        // toUInt().toLong()
+                        nextBits(bitCount).toLong() and 0xFFFF_FFFF
+                    }
+                    nHigh == 1 ->
+                        // toUInt().toLong()
+                        nextInt().toLong() and 0xFFFF_FFFF
+                    else -> {
+                        val bitCount = fastLog2(nHigh)
+                        nextBits(bitCount).toLong().shl(32) + nextInt()
+                    }
+                }
+            } else {
+                var v: Long
+                do {
+                    val bits = nextLong().ushr(1)
+                    v = bits % n
+                } while (bits - v + (n - 1) < 0)
+                rnd = v
+            }
+            return from + rnd
+        } else {
+            while (true) {
+                val rnd = nextLong()
+                if (rnd in from until until) return rnd
+            }
+        }
+    }
 
     /**
      * Gets the next random [Boolean] value.
      *
      * @sample samples.random.Randoms.nextBoolean
      */
-    public open fun nextBoolean(): Boolean
+    public open fun nextBoolean(): Boolean = nextBits(1) != 0
 
     /**
      * Gets the next random [Double] value uniformly distributed between 0 (inclusive) and 1 (exclusive).
      *
      * @sample samples.random.Randoms.nextDouble
      */
-    public open fun nextDouble(): Double
+    public open fun nextDouble(): Double = doubleFromParts(nextBits(26), nextBits(27))
 
     /**
      * Gets the next random non-negative `Double` from the random number generator less than the specified [until] bound.
@@ -117,7 +178,7 @@ public expect abstract class Random constructor() {
      *
      * @sample samples.random.Randoms.nextDoubleFromUntil
      */
-    public open fun nextDouble(until: Double): Double
+    public open fun nextDouble(until: Double): Double = nextDouble(0.0, until)
 
     /**
      * Gets the next random `Double` from the random number generator in the specified range.
@@ -130,14 +191,24 @@ public expect abstract class Random constructor() {
      *
      * @sample samples.random.Randoms.nextDoubleFromUntil
      */
-    public open fun nextDouble(from: Double, until: Double): Double
+    public open fun nextDouble(from: Double, until: Double): Double {
+        checkRangeBounds(from, until)
+        val size = until - from
+        val r = if (size.isInfinite() && from.isFinite() && until.isFinite()) {
+            val r1 = nextDouble() * (until / 2 - from / 2)
+            from + r1 + r1
+        } else {
+            from + nextDouble() * size
+        }
+        return if (r >= until) until.nextDown() else r
+    }
 
     /**
      * Gets the next random [Float] value uniformly distributed between 0 (inclusive) and 1 (exclusive).
      *
      * @sample samples.random.Randoms.nextFloat
      */
-    public open fun nextFloat(): Float
+    public open fun nextFloat(): Float = nextBits(24) / (1 shl 24).toFloat()
 
     /**
      * Fills a subrange of the specified byte [array] starting from [fromIndex] inclusive and ending [toIndex] exclusive
@@ -147,7 +218,30 @@ public expect abstract class Random constructor() {
      *
      * @sample samples.random.Randoms.nextBytes
      */
-    public open fun nextBytes(array: ByteArray, fromIndex: Int = 0, toIndex: Int = array.size): ByteArray
+    public open fun nextBytes(array: ByteArray, fromIndex: Int = 0, toIndex: Int = array.size): ByteArray {
+        require(fromIndex in 0..array.size && toIndex in 0..array.size) { "fromIndex ($fromIndex) or toIndex ($toIndex) are out of range: 0..${array.size}." }
+        require(fromIndex <= toIndex) { "fromIndex ($fromIndex) must be not greater than toIndex ($toIndex)." }
+
+        val steps = (toIndex - fromIndex) / 4
+
+        var position = fromIndex
+        repeat(steps) {
+            val v = nextInt()
+            array[position] = v.toByte()
+            array[position + 1] = v.ushr(8).toByte()
+            array[position + 2] = v.ushr(16).toByte()
+            array[position + 3] = v.ushr(24).toByte()
+            position += 4
+        }
+
+        val remainder = toIndex - position
+        val vr = nextBits(remainder * 8)
+        for (i in 0 until remainder) {
+            array[position + i] = vr.ushr(i * 8).toByte()
+        }
+
+        return array
+    }
 
     /**
      * Fills the specified byte [array] with random bytes and returns it.
@@ -156,14 +250,14 @@ public expect abstract class Random constructor() {
      *
      * @sample samples.random.Randoms.nextBytes
      */
-    public open fun nextBytes(array: ByteArray): ByteArray
+    public open fun nextBytes(array: ByteArray): ByteArray = nextBytes(array, 0, array.size)
 
     /**
      * Creates a byte array of the specified [size], filled with random bytes.
      *
      * @sample samples.random.Randoms.nextBytes
      */
-    public open fun nextBytes(size: Int): ByteArray
+    public open fun nextBytes(size: Int): ByteArray = nextBytes(ByteArray(size))
 
 
     /**
@@ -173,27 +267,36 @@ public expect abstract class Random constructor() {
      *
      * @sample samples.random.Randoms.defaultRandom
      */
-    companion object Default : Random {
-        override fun nextBits(bitCount: Int): Int
-        override fun nextInt(): Int
-        override fun nextInt(until: Int): Int
-        override fun nextInt(from: Int, until: Int): Int
+    companion object Default : Random(), Serializable {
+        private const val serialVersionUID = 6449517406788167720L
+        private val defaultRandom: Random = defaultPlatformRandom()
 
-        override fun nextLong(): Long
-        override fun nextLong(until: Long): Long
-        override fun nextLong(from: Long, until: Long): Long
+        private fun readResolve(): Any = Default
 
-        override fun nextBoolean(): Boolean
+        override fun nextBits(bitCount: Int): Int = defaultRandom.nextBits(bitCount)
+        override fun nextInt(): Int = defaultRandom.nextInt()
+        override fun nextInt(until: Int): Int = defaultRandom.nextInt(until)
+        override fun nextInt(from: Int, until: Int): Int = defaultRandom.nextInt(from, until)
 
-        override fun nextDouble(): Double
-        override fun nextDouble(until: Double): Double
-        override fun nextDouble(from: Double, until: Double): Double
+        override fun nextLong(): Long = defaultRandom.nextLong()
+        override fun nextLong(until: Long): Long = defaultRandom.nextLong(until)
+        override fun nextLong(from: Long, until: Long): Long = defaultRandom.nextLong(from, until)
 
-        override fun nextFloat(): Float
+        override fun nextBoolean(): Boolean = defaultRandom.nextBoolean()
 
-        override fun nextBytes(array: ByteArray): ByteArray
-        override fun nextBytes(size: Int): ByteArray
-        override fun nextBytes(array: ByteArray, fromIndex: Int, toIndex: Int): ByteArray
+        override fun nextDouble(): Double = defaultRandom.nextDouble()
+        override fun nextDouble(until: Double): Double = defaultRandom.nextDouble(until)
+        override fun nextDouble(from: Double, until: Double): Double = defaultRandom.nextDouble(from, until)
+
+        override fun nextFloat(): Float = defaultRandom.nextFloat()
+
+        override fun nextBytes(array: ByteArray): ByteArray = defaultRandom.nextBytes(array)
+        override fun nextBytes(size: Int): ByteArray = defaultRandom.nextBytes(size)
+        override fun nextBytes(array: ByteArray, fromIndex: Int, toIndex: Int): ByteArray = defaultRandom.nextBytes(
+            array,
+            fromIndex,
+            toIndex
+        )
     }
 }
 
@@ -228,6 +331,7 @@ public fun Random(seed: Int): Random = XorWowRandom(seed, seed.shr(31))
 public fun Random(seed: Long): Random = XorWowRandom(seed.toInt(), seed.shr(32).toInt())
 
 
+
 /**
  * Gets the next random `Int` from the random number generator in the specified [range].
  *
@@ -260,6 +364,9 @@ public fun Random.nextLong(range: LongRange): Long = when {
     else -> nextLong()
 }
 
+
+internal expect fun defaultPlatformRandom(): Random
+internal expect fun doubleFromParts(hi26: Int, low27: Int): Double
 
 @OptIn(ExperimentalStdlibApi::class)
 internal fun fastLog2(value: Int): Int = 31 - value.countLeadingZeroBits()
