+++ libraries/stdlib/src/kotlin/random/XorWowRandom.kt
@@ -15,16 +15,45 @@ package kotlin.random
  * Available at https://www.jstatsoft.org/v08/i14/paper
  *
  */
-internal expect class XorWowRandom internal constructor(
-    x: Int,
-    y: Int,
-    z: Int,
-    w: Int,
-    v: Int,
-    addend: Int,
-) : Random {
-    internal constructor(seed1: Int, seed2: Int)
+internal class XorWowRandom internal constructor(
+    private var x: Int,
+    private var y: Int,
+    private var z: Int,
+    private var w: Int,
+    private var v: Int,
+    private var addend: Int
+) : Random(), Serializable {
 
-    override fun nextInt(): Int
-    override fun nextBits(bitCount: Int): Int
+    internal constructor(seed1: Int, seed2: Int) :
+            this(seed1, seed2, 0, 0, seed1.inv(), (seed1 shl 10) xor (seed2 ushr 4))
+
+    init {
+        require((x or y or z or w or v) != 0) { "Initial state must have at least one non-zero element." }
+
+        // some trivial seeds can produce several values with zeroes in upper bits, so we discard first 64
+        repeat(64) { nextInt() }
+    }
+
+    override fun nextInt(): Int {
+        // Equivalent to the xorxow algorithm
+        // From Marsaglia, G. 2003. Xorshift RNGs. J. Statis. Soft. 8, 14, p. 5
+        var t = x
+        t = t xor (t ushr 2)
+        x = y
+        y = z
+        z = w
+        val v0 = v
+        w = v0
+        t = (t xor (t shl 1)) xor v0 xor (v0 shl 4)
+        v = t
+        addend += 362437
+        return t + addend
+    }
+
+    override fun nextBits(bitCount: Int): Int =
+        nextInt().takeUpperBits(bitCount)
+
+    companion object {
+        private const val serialVersionUID = -6943221717878108522L
+    }
 }
